% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsderivs.R
\name{bs.derivs}
\alias{bs.derivs}
\title{Derivative of B-spline}
\usage{
bs.derivs(x, derivs = 0, df = NULL, knots = NULL, degree = 3,
  intercept = FALSE, Boundary.knots = range(x))
}
\arguments{
\item{x}{object of B-splines}

\item{derivs}{a number of derivatives}

\item{df}{degrees of freedom of B-splines}

\item{knots}{a vector of internal knots}

\item{degree}{degrees of B-splines}

\item{intercept}{a logical vector}

\item{Boundary.knots}{a vector of boundary knots}
}
\value{
The function \code{bs.derivs} returns a component:
\item{resmat}{derivatives of B-spline}
}
\description{
takes the derivative of the B-splines
}
\details{
The function \code{bs.derivs} performs derivatives of B-splines
}
\author{
Giorgos Bakoyannis, \email{gbakogia at iu dot edu}

Jun Park, \email{jp84 at iu dot edu}
}
\keyword{bs.derivs}
