% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bssmle_se.R
\name{bssmle_se}
\alias{bssmle_se}
\title{Bootstrap varince-covariance estimation}
\usage{
bssmle_se(formula, data, nboot, alpha, do.par)
}
\arguments{
\item{formula}{the formula object relating survival object \code{Surv2(v, u, event)} to a set of covariates}

\item{data}{data frame to be used}

\item{nboot}{the number of bootstrap samples for estimating variances and covariances of the estimated regression coefficients. If \code{nboot = 0}, \code{ciregic} does dot perform bootstrap estimation of the variance matrix of the regression parameter estimates and returns \code{NA} in the place of the estimated variance matrix of the regression parameter estimates.}

\item{alpha}{\eqn{\alpha=(\alpha1, \alpha2)} contains parameters that that define the link functions from class of generalized odds-rate transformation models. The components \eqn{\alpha1} and \eqn{\alpha2} should both be \eqn{\ge 0}. If \eqn{\alpha1 = 0}, the user assumes a proportional subdistribution hazards or Fine-Gray model for cause of failure 1. If \eqn{\alpha2 = 1}, the user assumes a proportional odds model for cause of failure 2.}

\item{do.par}{using parallel computing for bootstrap. If \code{TRUE}, parallel computing will be used during the bootstrap estimation of the variance-covariance matrix for the regression parameter estimates.}
}
\value{
\code{bssmle_se} returns components
\item{numboot}{the number of bootstrap with successful convergence}
\item{Sigma}{the estimated bootstrap variance-covariance matrix of the estimated regression coefficients}
}
\description{
Bootstrap varince estimation for the estimated regression coefficients
}
\details{
\code{bssmle_se} is the function to estimate bootstrap standard errors for the estimated regression coefficients from the function \code{bssmle}.
}
\examples{
intccr:::bssmle_se(Surv2(v, u, c) ~ z1 + z2, data = simdat,
                   nboot = 1, alpha = c(1, 1), do.par = FALSE)
}
\keyword{bssmle_se}
