% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsderivs.R
\name{dbs}
\alias{dbs}
\title{Derivative of B-spline}
\usage{
dbs(
  x,
  derivs = 1L,
  df = NULL,
  knots = NULL,
  degree = 3L,
  intercept = FALSE,
  Boundary.knots = range(x, na.rm = TRUE)
)
}
\arguments{
\item{x}{object of B-splines}

\item{derivs}{a number of derivatives}

\item{df}{degrees of freedom of B-splines}

\item{knots}{a vector of internal knots}

\item{degree}{degrees of B-splines}

\item{intercept}{a logical vector}

\item{Boundary.knots}{a vector of boundary knots}
}
\value{
The function \code{dbs} returns a component:
\item{dMat}{B-spline matrix}
}
\description{
Generates the derivative of the B-splines basis matrix.
}
\details{
The function \code{dbs} performs derivatives of B-splines
}
\author{
Jun Park, \email{jun.park@alumni.iu.edu}

Giorgos Bakoyannis, \email{gbakogia@iu.edu}
}
\keyword{dbs}
