% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\docType{data}
\name{fit}
\alias{fit}
\title{Output of \code{ciregic}}
\format{A list of components.}
\usage{
fit
}
\description{
Object contains the output of the function \code{ciregic}. Standard errors were estimated by the least-squares method.
}
\examples{
fit
}
\keyword{output}
