% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_lt.R
\docType{data}
\name{fit_lt}
\alias{fit_lt}
\title{Output of \code{ciregic_lt}}
\format{
A list of components.
}
\usage{
fit_lt
}
\description{
Object contains the output of the function \code{ciregic_lt}. Standard errors were estimated by the least-squares method.
}
\examples{
fit_lt
}
\keyword{output}
