% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longdat.R
\docType{data}
\name{longdata}
\alias{longdata}
\title{Simulated data consisting of 200 individuals with 2 covariates.}
\format{A data frame with 868 rows and 5 variables.}
\usage{
longdata
}
\description{
Data containing subject id, a series of time points, cause of failure, and covariates.
}
\examples{
longdata
}
\keyword{datasets}
