% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longdata_lt.R
\docType{data}
\name{longdata_lt}
\alias{longdata_lt}
\title{Simulated left-truncated and interval-censored competing risks data - long format}
\format{
A data frame with 275 unique individuals and 6 variables.
}
\usage{
longdata_lt
}
\description{
Data containing observation time points, a left-truncation time, cause of failure, and baseline covariates with 275 observations.
}
\examples{
library(intccr)
data(longdata_lt)
}
\keyword{datasets}
