% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_b.R
\name{naive_b}
\alias{naive_b}
\title{naive_b}
\usage{
naive_b(data, v, u, c, q)
}
\arguments{
\item{data}{Data frame to be used.}

\item{v}{The last observation time prior to the failure}

\item{u}{The first observation time after the failure}

\item{c}{the event or censoring indicator. \code{event} should include 0, 1 or 2,
denoting right-censoring, failure from cause 1 and failure from cause 2, respectively.
If \code{event}=0 (i.e. right-censored observation) then \code{u} is not included in any calculation as it corresponds to \eqn{\infty}.}

\item{q}{The dimension of design matrix}
}
\description{
naive_b
}
\keyword{naive_b}
