% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsderivs.R
\name{predict.dbs}
\alias{predict.dbs}
\title{Prediction of derivative of B-spline}
\usage{
\method{predict}{dbs}(object, newx)
}
\arguments{
\item{object}{returned object of B-splines}

\item{newx}{a vector of points}
}
\value{
The function \code{predict} returns a predicted B-splies.
}
\description{
Evaluates the derivative of the B-splines basis matrix at given values.
}
\details{
The function \code{predict} is a generic function of \code{bs.derivs}
}
\author{
Giorgos Bakoyannis, \email{gbakogia@iu.edu}

Jun Park, \email{jp84@alumni.iu.edu}
}
\keyword{predict}
