% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoHIVlong.R
\docType{data}
\name{pseudo.HIV.long}
\alias{pseudo.HIV.long}
\title{Artificial HIV dataset}
\format{A data frame with 22710 rows and 6 variables.}
\usage{
pseudo.HIV.long
}
\description{
Artificial dataset that was simulated to resemble the HIV study on loss to HIV care and death in sub-Saharan Africa, that was presented in Bakoyannis, Yu, & Yiannoutsos (2017). It contains subject id, observation times, cause of failure, and covariates.
}
\examples{
head(pseudo.HIV.long, n = 20)

}
\references{
{Bakoyannis, G., Yu, M., and Yiannoutsos C. T. (2017). Semiparametric regression on cumulative incidence function with interval-censored competing risks data. \emph{Statistics in Medicine}, \strong{36}:3683-3707.}
}
\keyword{datasets}
