% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat.R
\docType{data}
\name{simdat}
\alias{simdat}
\title{A simulated dataset consisting of 200 interval-censored observations and 2 covariates.}
\format{A data frame with 200 rows and 5 variables.}
\usage{
data(simdat)
}
\description{
A dataset containing the last time prior to the event, the first time after the event, cause of failure, and covariates.
}
\examples{
simdat
}
\keyword{datasets}
