% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata_aipw.R
\docType{data}
\name{simdata_aipw}
\alias{simdata_aipw}
\title{Simulated interval censored data with 2 covariates in the presence of \eqn{50\%} of missing cause of failure - wide format}
\format{A data frame with 200 rows and 7 variables:
\describe{
  \item{id}{subject id}
  \item{v}{the last observation time prior to the event}
  \item{u}{the first observation time after the event}
  \item{c}{cause of failure with missing}
  \item{z1}{binary variable}
  \item{z2}{continuous variable}
  \item{a}{auxiliary variable}
}}
\usage{
simdata_aipw
}
\description{
The dataset containing the individual identification number, the last time prior to the event, the first time after the event, cause of failure with \eqn{50\%} of missingness, and covariates.
}
\examples{
library(intccr)
data(simdata_aipw)
}
\keyword{datasets}
