% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata_lt.R
\docType{data}
\name{simdata_lt}
\alias{simdata_lt}
\title{Simulated left-truncated and interval-censored competing risks data - wide format}
\format{A data frame with 275 unique individuals and 7 variables.}
\usage{
data(simdata_lt)
}
\description{
Data containing an individual identificatio number, a left-truncated time, the last and first observation time prior to the event and after the event, cause of failure, and baseline covariates with 275 observations.
}
\examples{
library(intccr)
data(simdata_lt)
}
\keyword{datasets}
