% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciregic_lt.R
\name{summary.ciregic_lt}
\alias{summary.ciregic_lt}
\title{Summary of \code{ciregic_lt}}
\usage{
\method{summary}{ciregic_lt}(object, ...)
}
\arguments{
\item{object}{an object of class \code{ciregic_lt}, which is a result of a call to \code{ciregic_lt}}

\item{...}{further arguments}
}
\value{
The function \code{\link[intccr]{summary.ciregic_lt}} returns a list of summary statistics of the model from \code{object}.
\item{varnames}{a vector containing variable names}
\item{coefficients}{a vector of the regression coefficient estimates}
\item{se}{a bootstrap standard error of the coefficients}
\item{z}{z value of the estimated coefficients}
\item{p}{p value of the estimated coefficients}
\item{call}{a matched call}
}
\description{
\code{summary} method for class \code{ciregic_lt}
}
\details{
The function \code{summary.ciregic_lt} returns the coefficients, bootstrap standard errors, and etc. Additionally, 'significance star' is included.
}
\examples{
## Continuing the ciregic_lt(...) example
sfit_lt <- summary(fit_lt)
sfit_lt
}
\seealso{
The fitted semiparametric regression on cumulative incidence function with interval-censored competing risks data \code{\link[intccr]{ciregic_lt}} and values of the predicted cumulative incidence functions \code{\link[intccr]{predict.ciregic_lt}}
}
