% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv2.R
\name{Surv2}
\alias{Surv2}
\title{Creating data frame}
\usage{
Surv2(v, u, event)
}
\arguments{
\item{v}{the last observation time prior to the failure; \eqn{0\le v \le u}.}

\item{u}{the first observation time after the failure; \eqn{u \ge 0}.}

\item{event}{an indicator of event type. If an observation is righ-censored, \code{event = 0}; otherwise, \code{event = 1} or \code{event = 2}, where \code{1} represents the event type 1, and \code{2} represents the event type 2. The current version of package only allows for two event types.}
}
\value{
data frame
}
\description{
The function \code{Surv2} generates the survival object to be treated as the response from \code{ciregic}.
}
\details{
The function \code{Surv2} provides a response data frame which is used in the function \code{ciregic}.
}
\examples{
attach(simdata)
Surv2(v, u, c)

}
\author{
Giorgos Bakoyannis, \email{gbakogia at iu dot edu}

Jun Park, \email{jp84 at iu dot edu}
}
\keyword{Surv2}
