% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciregic_lt.R
\name{vcov.ciregic_lt}
\alias{vcov.ciregic_lt}
\title{Variance-covariance matrix of \code{ciregic_lt}}
\usage{
\method{vcov}{ciregic_lt}(object, ...)
}
\arguments{
\item{object}{an object of class \code{ciregic_lt}, which is a result of a call to \code{ciregic_lt}}

\item{...}{further arguments}
}
\value{
The estimated bootstrap variance-covariance matrix
}
\description{
\code{vcov} method for class \code{ciregic_lt}
}
\details{
The function \code{vcov} returns the variance-covariance matrix of the fitted semiparametric regression model.
}
\examples{
## Continuing the ciregic_lt(...) example
vcov(fit_lt)
}
\seealso{
The fitted semiparametric regression on cumulative incidence function with interval-censored competing risks data \code{\link[intccr]{ciregic_lt}}, summary of the fitted semiparametric regression model \code{\link[intccr]{summary.ciregic_lt}}, and values of predicted cumulative incidence functions \code{\link[intccr]{predict.ciregic_lt}}
}
