% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bssmle.R
\name{bssmle}
\alias{bssmle}
\title{bssmle}
\usage{
bssmle(formula, data, alpha)
}
\arguments{
\item{formula}{A formula object relating survival object \code{Surv2(v, u, event)} to a set of covariates}

\item{data}{Data frame to be used.}

\item{alpha}{\eqn{\alpha=(\alpha1, \alpha2)} contains parameters that that define the link functions from class of generalized odds-rate transformation models. The components \eqn{\alpha1} and \eqn{\alpha2} should both be \eqn{\ge 0}. If \eqn{\alpha1 = 0} the user assumes a proportional subdistribution hazards or Fine-Gray model for cause of failure 1. If \eqn{\alpha2 = 1} the user assumes a proportional odds model for cause of failure 2.}
}
\description{
bssmle
}
\keyword{bssmle}
