% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bssmle_lse.R
\name{bssmle_lse}
\alias{bssmle_lse}
\title{Least-Squares Estimator of the Information Matrix}
\usage{
bssmle_lse(obj)
}
\arguments{
\item{obj}{a list of objectives from \code{bssmle}}
}
\value{
The function \code{bssmle_lse} returns a list of components:
\item{Sigma}{the estimated variance-covariance matrix for the estimated regression coefficients}
}
\description{
Performs the least-squares methods to estimate the information matrix for the estimated regression coefficients
}
\details{
The function \code{bssmle_lse} estimates the information matrix for the estimated regression coefficients from the function \code{bssmle} using the lease-squares method.
}
\references{
{Zhang, Y., Hua, L., and Huang, J. (2010), A spline-based semiparametric maximum likelihood estimation method for the Cox model with interval-censoed data. \emph{Scandinavian Journal of Statistics}, \strong{37}:338-354.}
}
\author{
Jun Park, \email{jp84 at iu dot edu}

Giorgos Bakoyannis, \email{gbakogia at iu dot edu}
}
\keyword{bssmle_lse}
