% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longdata.R
\docType{data}
\name{longdata}
\alias{longdata}
\title{Simulated interval-censored competing risks data - long format}
\format{A data frame with 868 rows and 5 variables.}
\usage{
longdata
}
\description{
The data containing the subject id, series of time points, cause of failure, and covariates with 200 observations.
}
\examples{
library(intccr)
data(longdata)
}
\keyword{datasets}
