% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longdata.R
\docType{data}
\name{longdata}
\alias{longdata}
\title{Simulated long format data}
\format{A data frame with 868 rows and 5 variables.}
\usage{
longdata
}
\description{
Data containing subject id, a series of time points, cause of failure, and covariates with 200 observations.
}
\examples{
head(longdata, n = 5)
}
\keyword{datasets}
