% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Simulated data consisting of 200 interval-censored observations with 2 covariates.}
\format{A data frame with 200 rows and 5 variables.}
\usage{
simdata
}
\description{
Data containing the last time point prior to the event, the first time point after the event, cause of failure, and covariates.
}
\examples{
simdata
}
\keyword{datasets}
