% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Simulated data}
\format{A data frame with 200 rows and 5 variables.}
\usage{
simdata
}
\description{
Data containing the last time point prior to the event, the first time point after the event, cause of failure, and covariates with 200 observations.
}
\examples{
head(simdata, n = 5)
}
\keyword{datasets}
