% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv2.R
\name{Surv2}
\alias{Surv2}
\title{Survival object}
\usage{
Surv2(v, u, event)
}
\arguments{
\item{v}{The last observation time prior to the failure; \eqn{0\le v \le u}.}

\item{u}{The first observation time after the failure; \eqn{u \ge 0}.}

\item{event}{The cause of failure indicator. If observation is righ-censored, \code{event=0}, otherwise \code{event= 1 or 2}, where \code{1} represents first cause of failure
, and \code{2} represents second cause of failure. The current version of package only allows for two causes of failure.}
}
\description{
\code{Surv2} generates survival object for a given cause of failure.
}
\keyword{Surv2}
